Moved 'page' components into scenes folder
