--- src/components/ConferenceNewPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ConferenceNewPage'
