--- src/components/ConferencePage/components/index.ts
@@ -1,1 +1,0 @@
-export { default as CFPHeader } from './CFPHeader'
