+++ src/scenes/ConferenceForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ConferenceForm'
