+++ src/scenes/ConferenceForm/utils/getConferenceData.ts
