+++ src/scenes/ConferencePage/ConferencePage.scss
