+++ src/scenes/ConferencePage/CurrentRefinement.scss
