+++ src/scenes/ConferencePage/components/CFPHeader/CFPHeader.scss
