+++ src/scenes/ConferencePage/components/CFPHeader/index.ts
