+++ src/scenes/ConferencePage/components/ConferenceItem/ConferenceItem.scss
