+++ src/scenes/ConferencePage/components/ConferenceItem/index.ts
