+++ src/scenes/ConferencePage/components/ConferenceItem/tests/ConferenceItem.test.ts
