+++ src/scenes/ConferencePage/components/ConferenceItem/utils.ts
