+++ src/scenes/ConferencePage/components/ConferenceList/ConferenceList.scss
