+++ src/scenes/ConferencePage/components/ConferenceList/index.ts
