+++ src/scenes/ConferencePage/components/index.ts
@@ -1,0 +1,3 @@
+export { default as CFPHeader } from './CFPHeader'
+export { default as ConferenceList } from './ConferenceList'
+export { default as ConferenceItem } from './ConferenceItem'
