+++ src/scenes/index.ts
@@ -1,0 +1,2 @@
+export { default as ConferenceForm } from './ConferenceForm'
+export { default as ConferencePage } from './ConferencePage'
